refactor search methods to new search module
